backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.10.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-dotenv==1.0.1
scikit-learn==1.6.0
numpy==2.2.0
pandas==2.2.3
python-multipart==0.0.19
httpx==0.28.1
